% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMICor.R
\name{buildMICor}
\alias{buildMICor}
\title{build model-implied correlation matrix}
\usage{
buildMICor(eqns, pt)
}
\arguments{
\item{eqns}{}

\item{pt}{}
}
\description{
build model-implied correlation matrix
}
\keyword{internal}
