% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMissingSyntax.R
\name{buildMissingSyntax}
\alias{buildMissingSyntax}
\title{build the missing data model syntax}
\usage{
buildMissingSyntax(pt)
}
\description{
build the missing data model syntax
}
\keyword{internal}
