% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildRestrictMat.R
\name{buildRestrictMat}
\alias{buildRestrictMat}
\title{Build the restriction matrix}
\usage{
buildRestrictMat(d)
}
\arguments{
\item{d}{A list containing equation information.}
}
\description{
Function returns a list \code{r} with information regarding
coefficient restrictions and estimation. The field \code{R} 
contains a restriction matrix for imposing cross-coefficient
restrictions. If no cross-coefficient restrictions are present
\code{R} is set to \code{NULL}. The vector \code{q} contains
a vector of lagrangean multipliers. The vector \code{constrained} 
gives the constrained coefficients in convenient form.
}
\details{
If an equation contains any categorical variables or variables
with missing values in addition to a restricted coefficient,
an error is thrown.
}
\keyword{internal}
