% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildVarCovSyntax.R
\name{buildVarCovSyntax}
\alias{buildVarCovSyntax}
\title{build the variance covariance point estimates model syntax}
\usage{
buildVarCovSyntax(pt)
}
\description{
build the variance covariance point estimates model syntax
}
\keyword{internal}
