% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est2SLSCoef.R
\name{est2SLSCoef}
\alias{est2SLSCoef}
\title{estimate the 2SLS point estimates}
\usage{
est2SLSCoef(d = d, r = NULL, poly.mat = NULL, sscp.mat = NULL)
}
\description{
estimate the 2SLS point estimates
}
\keyword{internal}
