% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est2SLSCoefCov.R
\name{est2SLSCoefCov}
\alias{est2SLSCoefCov}
\title{estimate the 2SLS coefficient covariance matrix}
\usage{
est2SLSCoefCov(
  d,
  poly.mat = NULL,
  cov.mat = NULL,
  mean.vec = NULL,
  acov = NULL,
  acov.sat = NULL,
  r = NULL
)
}
\description{
estimate the 2SLS coefficient covariance matrix
}
\keyword{internal}
