% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est2SLSSigmaSq.R
\name{est2SLSSigmaSq}
\alias{est2SLSSigmaSq}
\title{estimate the 2SLS equation specific sigma squared}
\usage{
est2SLSSigmaSq(d, cov.mat)
}
\description{
estimate the 2SLS equation specific sigma squared
}
\keyword{internal}
