% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estResidualCov.R
\name{estResidualCov}
\alias{estResidualCov}
\title{Calculate the residual covariance matrix}
\usage{
estResidualCov(d, sample.cov)
}
\description{
Calculate the residual covariance matrix
}
\keyword{internal}
