% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estVarCovar.R
\name{estVarCovar}
\alias{estVarCovar}
\title{estimate the variance and covariance parameters}
\usage{
estVarCovar(data, g, eqns, pt, ordered, se, missing)
}
\description{
estimate the variance and covariance parameters
}
\keyword{internal}
