% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estVarCovarCoef.R
\name{estVarCovarCoef}
\alias{estVarCovarCoef}
\title{estimate the variance and covariance parameters}
\usage{
estVarCovarCoef(data, g, eqns, pt, ordered, vc.coef.estimator = "ML",
  vc.coef.missing = "FIML")
}
\description{
estimate the variance and covariance parameters
}
\keyword{internal}
