% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estVarCovarCoefs.R
\name{estVarCovarCoefs}
\alias{estVarCovarCoefs}
\title{estimate the variance and covariance parameters}
\usage{
estVarCovarCoefs(data, g, eqns, pt, ordered)
}
\description{
estimate the variance and covariance parameters
}
\keyword{internal}
