% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorScores.R
\name{factorScores}
\alias{factorScores}
\title{provides factor scores for miiv model}
\usage{
factorScores(model, data, lv_scores, lv_estimator)
}
\description{
provides factor scores for miiv model
}
\keyword{internal}
