% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModelList.R
\name{getModelList}
\alias{getModelList}
\title{provides model lists for latent variable gimme}
\usage{
getModelList(
  data,
  model,
  scaling = "first.indicator",
  lag = FALSE,
  dfm = FALSE
)
}
\description{
provides model lists for latent variable gimme
}
\keyword{internal}
