% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miive.2sls.R
\name{miive.2sls.system}
\alias{miive.2sls.system}
\title{Two-stage least square estimator for a system of equations}
\usage{
miive.2sls.system(d, sample.cov, sample.mean, sample.nobs, se, restrictions)
}
\arguments{
\item{d}{a system of MIIV equation}

\item{sample.cov}{sample covariance matrix}

\item{sample.nobs}{number of observations}

\item{se}{should variance covariance matrix of the estimates be calculated}

\item{restrictions}{any equality constraints to be used in estimation}
}
\description{
Two-stage least square estimator for a system of equations
}
\keyword{internal}

