% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miive.piv.R
\name{miive.piv}
\alias{miive.piv}
\title{Polychoric instrumental variable (PIV) estimator for a system of equations}
\usage{
miive.piv(d, data, sample.cov, sample.mean, sample.nobs, est.only, restrictions,
  piv.opts = piv.opts)
}
\arguments{
\item{d}{a list containing the system of MIIV estimating equations}

\item{sample.cov}{sample covariance matrix}

\item{sample.nobs}{number of observations}

\item{est.only}{should we only calculate coefficient estimates}

\item{restrictions}{any equality constraints to be used in estimation}

\item{piv.opts}{Options to pass to lavaan's \code{\link[lavCor]{lavCor}} function.}
}
\description{
Polychoric instrumental variable (PIV) estimator for a system of equations
}
\keyword{internal}
