% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miivs.R
\name{miivs}
\alias{miivs}
\title{Model-implied instrumental variable (MIIV) search}
\usage{
miivs(model)
}
\arguments{
\item{model}{A model specified using lavaan model syntax. See the \code{model} 
argument within the \code{\link[lavaan]{lavaanify}} function for more information.
A model specified using lavaan model syntax. See the \code{model} argument within 
the \code{\link[lavaan]{lavaanify}} function for more information. See the
documentation below for a description of how to specify the scaling 
indicator in latent variable models and impose equality constraints on the
parameter estimates.}
}
\value{
A list of model equations.
}
\description{
A key step in the MIIV-2SLS approach is to transform the SEM by replacing the 
latent variables with their scaling indicators minus their errors.  Upon 
substitution the SEM is transformed from a model with latent variables to 
one containing observed variables with composite errors.  The miivs function 
automatically makes this transformation.
}
\details{
The \code{miivs} function displays a table containing the following 
information for each equation in the system:

\itemize{
 \item \code{LHS} The "dependent" variable.
 \item \code{RHS} The right hand side variables of the transformed equation.
 \item \code{Composite Disturbance}  Elements of the composite errors in the transformed equation.
 \item \code{MIIVs} The model-implied instrumental variables for each equation.
}
}
\section{Scaling Indicators}{

Following the \code{lavaan} model syntax, latent variables are defined 
using the \code{"=~"} operator.  For first order factors, the scaling indicator 
chosen is the first observed variable on the RHS of an equation. For the model 
below \code{Z1} would be chosen as the scaling indicator for \code{L1} and  
\code{Z4} would be chosen as the scaling indicator for \code{L2}.

\preformatted{model <- '
    L1 =~ Z1 + Z2 + Z3
    L2 =~ Z4 + Z5 + Z6
'
}
}

\section{Higher-order Factor Models}{

For example, in the model below, the  scaling indicator for the higher-order 
factor \code{H1} is taken to be \code{Z1}, the scaling indicator 
that would have been assigned to the first lower-order factor \code{L1}.


\preformatted{model <- '
    H1 =~ L1 + L2 
    L1 =~ Z1 + Z2 + Z3
    L2 =~ Z4 + Z5 + Z6
'
}
}

\examples{
bollen1989a_model <- '

    Eta1 =~ y1 + y2  + y3  + y4  
    Eta2 =~ y5 + y6  + y7  + y8    
    Xi1  =~ x1 + x2 + x3 

    Eta1 ~ Xi1  
    Eta2 ~ Xi1 
    Eta2 ~ Eta1 

    y1   ~~ y5
    y2   ~~ y4
    y2   ~~ y6
    y3   ~~ y7
    y4   ~~ y8
    y6   ~~ y8 
  '

miivs(bollen1989a_model)
}
\references{
Bollen, K. A. 1996.	An	Alternative	2SLS Estimator	for	Latent	
Variable	Models.	\emph{Psychometrika}, 61, 109-121.
}
\seealso{
\code{\link{miive}}
}
