% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIIVsem.R
\docType{package}
\name{MIIVsem-package}
\alias{MIIVsem}
\alias{MIIVsem-package}
\title{Model Implied Instrumental Variable Estimation of Structural Equation Models}
\description{
\pkg{MIIVsem} is a package for estimating structural equation models 
using model-implied instrumental variables (MIIVs).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zackfisher/MIIVsem}
  \item Report bugs at \url{https://github.com/zackfisher/MIIVsem/issues}
}

}
\author{
\strong{Maintainer}: Zachary Fisher \email{fish.zachary@gmail.com}

Authors:
\itemize{
  \item Kenneth Bollen
  \item Kathleen Gates
  \item Mikko Rönkkö
}

}
