% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minEigenStat.R
\name{minEigenStat}
\alias{minEigenStat}
\title{Calculate the minimum eigenvalue statistic of Cragg and Donald (1993)}
\usage{
minEigenStat(IVobs, MIIVs, data)
}
\arguments{
\item{IVobs}{a vector of endogenous RHS variables.}

\item{MIIVs}{a vector of instrumental variables.}

\item{data}{matrix or data.frame containing raw data.}
}
\description{
Calculate the minimum eigenvalue statistic of Cragg and Donald (1993)
}
\keyword{internal}
