% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.miive.R
\name{print.miive}
\alias{print.miive}
\title{Print method for a MIIV estimation object}
\usage{
\method{print}{miive}(x, ...)
}
\arguments{
\item{x}{a miive object}

\item{...}{Optional arguments to print, not used by user.}
}
\description{
Print method for a MIIV estimation object
}
