% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.miivs.R
\name{print.miivs}
\alias{print.miivs}
\title{Print method for a MIIV search object}
\usage{
\method{print}{miivs}(x, ...)
}
\arguments{
\item{x}{a miivs object}

\item{...}{Optional arguments to print, not used by user.}
}
\description{
Print method for a MIIV search object
}
