% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{processData}
\alias{processData}
\title{Process data.}
\usage{
processData(
  data = data,
  sample.cov = sample.cov,
  sample.mean = sample.mean,
  sample.nobs = sample.nobs,
  ordered = ordered,
  missing = missing,
  se = se,
  pt = pt
)
}
\description{
Process data.
}
\keyword{internal}
