% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruneExcessMIIVs.R
\name{pruneExcessMIIVs}
\alias{pruneExcessMIIVs}
\title{Prune MIIVs based on some specified criterion.}
\usage{
pruneExcessMIIVs(
  d,
  overid.degree,
  overid.method,
  data = NULL,
  sample.cov = NULL,
  sample.polychoric = NULL,
  sample.mean = NULL,
  sample.nobs = NULL,
  overid.boot = FALSE,
  cat.vars = NULL
)
}
\description{
Prune MIIVs based on some specified criterion.
}
\keyword{internal}
