% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrict.tests.R
\name{restrict.tests}
\alias{restrict.tests}
\title{F and Wald test statistics for a set of linear restrictions.}
\usage{
restrict.tests(x)
}
\arguments{
\item{x}{A fitted miive object contraining the model where
at least one restriction is imposted. See the 
\code{\link{miive}} function documentation for
how to impose restrictions using the model 
syntax.}
}
\value{
A list containing the following elements:

\tabular{ll}{
\code{f.test}\tab Wald test statistic\cr
\code{f.df}\tab Degrees of freedom for wald test\cr
\code{f.p}\tab \cr
\code{wald.test}\tab Wald test statistic\cr
\code{wald.df}\tab Degrees of freedom for wald test\cr
\code{wald.p}\tab P-value associated with Wald test.\cr
}
}
\description{
F and Wald test statistics for a set of linear restrictions.
}
\details{
The internal restrict.tests function provides two test 
statistics for a large-sample wald test of any linaer 
restrictions imposed on the MIIV-2SLS coefficient matrix.
The first statistic is an approximate F and  the second 
is Chi-square. Assumptions and  additional details for each
test are given by Greene (2003, p. 346-347).
}
\references{
Greene, W. H. (2000). Econometric analysis. Upper Saddle River, N.J: 
Prentice Hall.
}
\seealso{
\link{MIIVsem}{miivs}
}
\keyword{2sls}
\keyword{MIIV}
\keyword{MIIV-2SLS}
\keyword{PIV}
\keyword{SEM}
\keyword{instrument}
\keyword{internal}
\keyword{least-squares}
\keyword{tsls}
\keyword{two-stage}
