% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheasRSq.R
\name{sheasRSq}
\alias{sheasRSq}
\title{Calculate Shea's Partial R-squares}
\usage{
sheasRSq(IVobs, MIIVs, sample.cov, sample.mean, sample.nobs)
}
\arguments{
\item{IVobs}{a vector of endogenous RHS variables.}

\item{MIIVs}{a vector of instrumental variables.}

\item{sample.cov}{Numeric matrix. A sample variance-covariance 
matrix. The rownames and colnames attributes must contain all
the observed variable names indicated in the model syntax.}

\item{sample.mean}{A sample mean vector. If \code{sample.cov} is provided
and the \code{sample.mean} argument is \code{NULL}, intercepts
for all endogenous variables will not be estimated.}

\item{sample.nobs}{Number of observations in the full data frame.}
}
\description{
Calculate Shea's Partial R-squares
}
\keyword{internal}
