% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.miivs.R
\name{summary.miivs}
\alias{summary.miivs}
\title{Summary information for a MIIV search object}
\usage{
\method{summary}{miivs}(object, miivs.out = FALSE, eq.info = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{miivs}}

\item{miivs.out}{A logical indicating whether the model-implied 
instrumental variables found for \code{model} should be printed
to the console. This is a temporary convenience function to 
provide an editable, baseline format, for the \code{instruments}
argument of \code{\link{miive}}.}

\item{eq.info}{A logical indicating whether equation-specific 
information should be printed. Useful in models with a large number 
of variables.}

\item{...}{Optional arguments to summary, not used by user.}
}
\description{
Summary information for a MIIV search object
}
