% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{wald}
\alias{wald}
\title{wald test of set of linear restrictions}
\usage{
wald(x)
}
\arguments{
\item{x}{A fitted miive object contraining restricted coefficients}
}
\description{
wald test of set of linear restrictions
}
\keyword{internal}
