% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildMICor.R
\name{buildMICor}
\alias{buildMICor}
\title{Build model-implied correlation matrix}
\usage{
buildMICor(eqns, pt)
}
\arguments{
\item{pt}{}

\item{d}{}
}
\description{
Build model-implied correlation matrix
}
\keyword{internal}
