% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatesTable.R
\name{estimatesTable}
\alias{estimatesTable}
\title{return a dataframe of parameter estimates for a fitted model.}
\usage{
estimatesTable(x, v = NULL, sarg = FALSE)
}
\arguments{
\item{x}{An object of class miive}

\item{v}{A list containing variance snd covariance parameter information.}

\item{sarg}{Logical. Should Sargan test results be included.}
}
\description{
return a dataframe of parameter estimates for a fitted model.
}
