% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estVarCovar.R
\name{estVarCovar}
\alias{estVarCovar}
\title{Estimate the variance and covariance parameters}
\usage{
estVarCovar(data, d, d.un, pt, ordered)
}
\arguments{
\item{ordered}{}
}
\description{
Estimate the variance and covariance parameters
}
\keyword{internal}
