% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miive.2sls.R
\name{miive.2sls}
\alias{miive.2sls}
\title{Two-stage least square estimator for a system of equations}
\usage{
miive.2sls(d, g, r, est.only)
}
\arguments{
\item{d}{a list containing equation information}

\item{g}{a list containing data information}

\item{r}{a list containing coefficient restrictions}

\item{est.only}{should we only calculate coefficient estimates}
}
\description{
Two-stage least square estimator for a system of equations
}
\details{
MIIV estimation using estimation functions. An 
estimation function retuns a list containing the
following elements:

coefficients - a vector of estimated coefficients
coefCov      - variance covariance matrix of the estimates
               (optional, depending on the est.only argument)
residCov     - variance covariance matrix of the equation
               disturbances (optional, depending on the 
               est.only argument).
eqn -          a list of miiv equations and equation level
               estimation results. fields added include:
                 coefficients: coefficients
                 sigma       : equation sigma^2
                 sargan      : Sargan's test statistic
                 sargan.df   : df freedom for Sargan's
}
\keyword{internal}
