% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miive.2sls.R
\name{miive.2sls}
\alias{miive.2sls}
\title{Two-stage least square estimator for a system of equations}
\usage{
miive.2sls(d, data, sample.cov, sample.mean, sample.nobs, est.only,
  restrictions)
}
\arguments{
\item{d}{a list containing the system of MIIV estimating equations}

\item{sample.cov}{sample covariance matrix}

\item{sample.nobs}{number of observations}

\item{est.only}{should we only calculate coefficient estimates}

\item{restrictions}{any equality constraints to be used in estimation}
}
\description{
Two-stage least square estimator for a system of equations
}
\keyword{internal}
