% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miive.2sls.R
\name{miive.2sls}
\alias{miive.2sls}
\title{Two-stage least square estimator for a system of equations}
\usage{
miive.2sls(d, data, sample.cov, sample.mean, sample.nobs, est.only, piv.opts,
  factorIndex)
}
\arguments{
\item{d}{a list containing the system of MIIV estimating equations}

\item{sample.cov}{sample covariance matrix}

\item{sample.nobs}{number of observations}

\item{est.only}{should we only calculate coefficient estimates}

\item{piv.opts}{Options to pass to lavaan's \code{\link[lavCor]{lavCor}} function.}

\item{factorIndex}{logical vector indexing whether variables are factors}

\item{restrictions}{any equality constraints to be used in estimation}
}
\description{
Two-stage least square estimator for a system of equations
}
\keyword{internal}
