% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miive.2sls.R
\name{miive.2sls}
\alias{miive.2sls}
\title{two-stage least square estimator for a system of equations}
\usage{
miive.2sls(d, g, r, est.only, se, missing, var.cov, sarg.adjust = "none")
}
\arguments{
\item{d}{a list containing equation information}

\item{g}{a list containing data matrices and characteristics}

\item{r}{a list containing coefficient restrictions}

\item{est.only}{should we only calculate coefficient estimates}

\item{se}{se estimation}
}
\description{
two-stage least square estimator for a system of equations
}
\keyword{internal}
