% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miive.R
\name{miive}
\alias{miive}
\title{Model-implied instrumental variable (MIIV) estimation}
\usage{
miive(model = model, data = NULL, instruments = NULL, sample.cov = NULL,
  sample.mean = NULL, sample.nobs = NULL, sample.cov.rescale = TRUE,
  estimator = "2SLS", se = "standard", bootstrap = 1000L,
  est.only = FALSE, var.cov = FALSE, miiv.check = TRUE, ordered = NULL)
}
\arguments{
\item{model}{A model specified using lavaan model syntax or a \code{\link{miivs}} object. See the \code{model} argument within the \code{\link[lavaan]{lavaanify}} function for more information.}

\item{data}{A data frame, list or environment or an object coercible by \code{as.data.frame} to data frame.}

\item{instruments}{A user-supplied list of valid MIIVs for each equation. See Example 2 below.}

\item{sample.cov}{Numeric matrix. A sample variance-covariance matrix. The rownames and colnames must contain the observed variable names.}

\item{sample.mean}{A sample mean vector.}

\item{sample.nobs}{Number of observations in the full data frame.}

\item{sample.cov.rescale}{If \code{TRUE}, the sample covariance matrix provided by the user is internally rescaled by multiplying it with a factor (N-1)/N.}

\item{estimator}{Options \code{"2SLS"} or \code{"GMM"} for estimating the model parameters. Default is \code{"2SLS"}.}

\item{se}{If "standard", conventional closed form standard errors are computed. If "boot" or "bootstrap", bootstrap standard errors are computed using standard bootstrapping.}

\item{bootstrap}{Number of bootstrap draws, if bootstrapping is used.}

\item{est.only}{If \code{TRUE}, only the coefficients are returned.}

\item{var.cov}{If \code{TRUE}, variance and covariance parameters are estimated.}

\item{ordered}{A vector of variable names to be treated as ordered factors in generating the polychoric correlation matrix.}

\item{miivs.check}{Options to turn off check for user-supplied instruments validity as MIIVs.}
}
\value{
A list of class \code{miive} containing the following elements:

\tabular{ll}{
\code{coefficients}\tab A named vector of parameter estimates\cr
\code{coefCov}\tab A variance-covariance matrix of the parameter estimates\cr
\code{residCov}\tab A residual variance-covariance matrix\cr
\code{eqn}\tab Equation level estimation resutls and statistics\cr
\code{call}\tab The matched call\cr
}
}
\description{
Estimate SEM models using model-implied instrumental variables (MIIVs).
}
\details{
\itemize{
\item{\code{instruments}} {
  Using the \code{instruments} option you can specify the MIIVs directly 
  for each equation in the model. To utilize this option you must first 
  define a list of instruments using the syntax displayed below. After 
  the list is defined, set the \code{instruments} argument equal to 
  the name of the list of MIIVs. Note, \code{instruments} are specified 
  for an equation, and not for a specific endogenous variable.
  }
\item{\code{se}} {
  When \code{se} is set to \code{"boot"} or \code{"bootstrap"} standard errors
  are computed using the pairs bootstrap. These standard errors are based
  on the standard deviation of successful bootstrap replications.  The 
  \code{z-value} and \code{P(>|z|)} assume the ratio of the coefficient 
  estimate to the bootstrap standard deviation approximates a normal 
  distribution.  Note, the Sargan test statistic is calculated from the original 
  sample and is not a bootstrap estimate.
  }
}
}
\examples{
# Example 1

bollen1989a_model <- '

    Eta1 =~ y1 + y2  + y3  + y4  
    Eta2 =~ y5 + y6  + y7  + y8    
    Xi1  =~ x1 + x2 + x3 

    Eta1 ~ Xi1  
    Eta2 ~ Xi1 
    Eta2 ~ Eta1 

    y1   ~~ y5
    y2   ~~ y4
    y2   ~~ y6
    y3   ~~ y7
    y4   ~~ y8
    y6   ~~ y8 
  '

miive(model = bollen1989a_model, data = bollen1989a)
# Example 2

bollen1989a_model <- '

    Eta1 =~ y1 + y2  + y3  + y4  
    Eta2 =~ y5 + y6  + y7  + y8    
    Xi1  =~ x1 + x2 + x3 

    Eta1 ~ Xi1  
    Eta2 ~ Xi1 
    Eta2 ~ Eta1 

    y1   ~~ y5
    y2   ~~ y4
    y2   ~~ y6
    y3   ~~ y7
    y4   ~~ y8
    y6   ~~ y8 
'

my_instruments <- ' 
    y1 ~ x2 + x3                            
    y5 ~ y2 + y3 + y4 + x2              
    y2 ~ y3 + y7 + y8 + x2           
    y3 ~ y2 + y4 + y6 + y8        
    y4 ~ y3 + y6           
    y6 ~ y3 + y4 + y7 + x2            
    y7 ~ y2 + y4 + y6 + y8       
    y8 ~ y2 + y3 + y7 + x2          
    x2 ~ y1 + y5 + y2 + y3 + y4 + y6
    x3 ~ y1 + y5 + y2 + y3 + y4 + y6
'

miive(model = bollen1989a_model, 
      data = bollen1989a, 
      instruments = my_instruments)
# Example 3

bollen1989a_model_r <- '

    Eta1 =~ y1 + l2*y2  + l3*y3  + l4*y4  
    Eta2 =~ y5 + l2*y6  + l3*y7  + l4*y8    
    Xi1  =~ x1 + x2 + 0.5*x3 

    Eta1 ~ Xi1  
    Eta2 ~ Xi1 
    Eta2 ~ Eta1 

    y1   ~~ y5
    y2   ~~ y4
    y2   ~~ y6
    y3   ~~ y7
    y4   ~~ y8
    y6   ~~ y8

'

miive(model = bollen1989a_model_r, data = bollen1989a)
}
\references{
Bollen, K. A. 1996.	An	Alternative	2SLS Estimator	for	Latent	
Variable	Models.	\emph{Psychometrika}, 61, 109-121.

Bollen,	K. A. 2001.	Two-stage	Least	Squares	and	Latent	Variable	Models:	
Simultaneous	Estimation	and	Robustness	to	Misspecifications.
In	R.	Cudeck,	S.	Du	Toit,	and	D.	Sorbom	(Eds.),	Structural	
Equation	Modeling:	Present	and	Future,	A	Festschrift	in	Honor	of	Karl	
Joreskog	(pp. 119-138).	Lincoln,	IL: Scientific	Software.
}
\seealso{
\link{MIIVsem}{miivs}
}
