% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIIVsem.R
\docType{package}
\name{MIIVsem}
\alias{MIIVsem}
\alias{MIIVsem-package}
\title{In-line documentation for R.}
\description{
MIIVsem is a package for estimating structural equation models 
using model-implied instrumental variables (MIIVs).
}
\author{
Zachary Fisher, \cr
Ken Bollen, \cr
Kathleen Gates, \cr
Mikko Rönkkö \cr

Maintainer: Zachary Fisher
}
\seealso{
See \code{vignette("MIIVsem", package = "MIIVsem")} for an overview
  of the package, \code{vignette("miivs", package = "MIIVsem")} for model-
  implied instrumental variable search, and \code{vignette("miive", package = "MIIVsem")} 
  for estimating models.
}
\keyword{2SLS,}
\keyword{MIIV,}
\keyword{SEM}
\keyword{instrument,}
\keyword{least}
\keyword{squares,}
\keyword{stage}
\keyword{two}

