% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseInstrumentSyntax.R
\name{parseInstrumentSyntax}
\alias{parseInstrumentSyntax}
\title{Parse the instrument information}
\usage{
parseInstrumentSyntax(d, instruments, miiv.check)
}
\arguments{
\item{d}{A list containing equation information}

\item{instruments}{A list of user-supplied instruments. By defaults, \code{instruments} is set to \code{NULL}}

\item{miiv.check}{Logical indicating whether the validity of MIIVs should be checked}
}
\description{
Check to determine if (1) the user-supplied dependent variables exist in the
}
\details{
\describe{
  \item{\code{instruments}}{ 
    Using the \code{instruments} option you can specify the MIIVs directly 
    for each equation in the model. To utilize this option you must first 
    define a list of instruments using the syntax displayed below. After 
    the list is defined, set the \code{instruments} argument equal to the 
    name of the list of MIIVs. Note, \code{instruments} are specified for 
    an equation, and not for a specific endogenous variable. Also, the
    \code{instruments} argument can be used to prune the estimated model.
    For example, if the \code{instruments} argument is provided only those
    equationss whose dependent variable is specified will be estimated.
  }
  \item{\code{miiv.check}}{ 
    The default is true.\code{miiv.check} can be used to turn off MIIV 
    validity cehcks, for example, if instruments are not in model.
  }
}
}
\examples{

\dontrun{

  #-----------------------------------------------------#
  # Examples:                                           #
  #  Ex.1 (Valid miivs provided)                        #
  #-----------------------------------------------------#

  model <- '
     Eta1 =~ y1 + y2  + y3  + y4  
     Eta2 =~ y5 + y6  + y7  + y8    
     Xi1  =~ x1 + x2 + x3 

     Eta1 ~ Xi1  
     Eta2 ~ Xi1 
     Eta2 ~ Eta1 

     y1   ~~ y5
     y2   ~~ y4
     y2   ~~ y6
     y3   ~~ y7
     y4   ~~ y8
     y6   ~~ y8 
  '
  
  good_instruments <- ' 
    y1 ~ x2 + x3                            
    y5 ~ y2 + y3 + y4 + x2                
    y2 ~ y3 + y7 + y8 + x2           
    y3 ~ y2 + y4 + y6 + y8        
    y4 ~ y3 + y6           
    y6 ~ y3 + y4 + y7 + x2            
    y7 ~ y2 + y4 + y6 + y8       
    y8 ~ y2 + y3 + y7 + x2          
    x2 ~ y1 + y5 + y2 + y3 + y4 + y6
    x3 ~ y1 + y5 + y2 + y3 + y4 + y6
  '
  
  bad_instruments <- ' 
    y1 ~ x2 + x3 + y4                         
    y5 ~ y2 + y3 + y4 + x2                
    y2 ~ y3 + y7 + y8 + x2           
    y3 ~ y2 + y4 + y6 + y8        
    y4 ~ y3 + y6           
    y6 ~ y3 + y4 + y7 + x2            
    y7 ~ y2 + y4 + y6 + y8       
    y8 ~ y2 + y3 + y7 + x2          
    x2 ~ y1 + y5 + y2 + y3 + y4 + y6
    x3 ~ y1 + y5 + y2 + y3 + y4 + y6
  '

}
  
}
\keyword{internal}
