% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseInstrumentSyntax.R
\name{parseInstrumentSyntax}
\alias{parseInstrumentSyntax}
\title{Parse user-implied instrumental variables}
\usage{
parseInstrumentSyntax(d, instruments, miiv.check)
}
\description{
Parse user-implied instrumental variables
}
\keyword{internal}
