% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wald.R
\name{wald}
\alias{wald}
\title{F and Wald test statistics for a set of linear restrictions.}
\usage{
wald(x)
}
\arguments{
\item{x}{A fitted miive object contraining the model where
at least one restriction is imposted. See the 
\code{\link{miive}} function documentation for
how to impose restrictions using the model 
syntax.}
}
\value{
A list containing the following elements:

\tabular{ll}{
\code{wald.test}\tab Wald test statistic\cr
\code{wald.df}\tab Degrees of freedom for wald test\cr
\code{wald.p}\tab P-value associated with Wald test.\cr
\code{f.test}\tab Wald test statistic\cr
\code{f.df}\tab Degrees of freedom for wald test\cr
\code{f.p}\tab \cr
}
}
\description{
F and Wald test statistics for a set of linear restrictions.
}
\details{

}
\examples{
# Example 1

bollen1989a_model <- '

    Eta1 =~ y1 + y2  + y3  + y4  
    Eta2 =~ y5 + y6  + y7  + y8    
    Xi1  =~ x1 + x2 + x3 

    Eta1 ~ Xi1  
    Eta2 ~ Xi1 
    Eta2 ~ Eta1 

    y1   ~~ y5
    y2   ~~ y4
    y2   ~~ y6
    y3   ~~ y7
    y4   ~~ y8
    y6   ~~ y8 

  '


# Example 2

bollen1989a_model <- '

    Eta1 =~ y1 + y2  + y3  + y4  
    Eta2 =~ y5 + y6  + y7  + y8    
    Xi1  =~ x1 + x2 + x3 

    Eta1 ~ Xi1  
    Eta2 ~ Xi1 
    Eta2 ~ Eta1 

    y1   ~~ y5
    y2   ~~ y4
    y2   ~~ y6
    y3   ~~ y7
    y4   ~~ y8
    y6   ~~ y8 
'

my_instruments <- ' 
    y1 ~ x2 + x3                            
    y5 ~ y2 + y3 + y4 + x2              
    y2 ~ y3 + y7 + y8 + x2           
    y3 ~ y2 + y4 + y6 + y8        
    y4 ~ y3 + y6           
    y6 ~ y3 + y4 + y7 + x2            
    y7 ~ y2 + y4 + y6 + y8       
    y8 ~ y2 + y3 + y7 + x2          
    x2 ~ y1 + y5 + y2 + y3 + y4 + y6
    x3 ~ y1 + y5 + y2 + y3 + y4 + y6
'

miive(model = bollen1989a_model, 
      data = bollen1989a, 
      instruments = my_instruments)
# Example 3

bollen1989a_model_r <- '

    Eta1 =~ y1 + l2*y2  + l3*y3  + l4*y4  
    Eta2 =~ y5 + l2*y6  + l3*y7  + l4*y8    
    Xi1  =~ x1 + x2 + 0.5*x3 

    Eta1 ~ Xi1  
    Eta2 ~ Xi1 
    Eta2 ~ Eta1 

    y1   ~~ y5
    y2   ~~ y4
    y2   ~~ y6
    y3   ~~ y7
    y4   ~~ y8
    y6   ~~ y8

'

miive(model = bollen1989a_model_r, data = bollen1989a)
}
\references{
#' Greene, W. H. (2000). Econometric analysis. Upper Saddle River, N.J: 
Prentice Hall.
}
\seealso{
\link{MIIVsem}{miivs}
}
\keyword{2sls}
\keyword{MIIV}
\keyword{MIIV-2SLS}
\keyword{PIV}
\keyword{SEM}
\keyword{instrument}
\keyword{internal}
\keyword{least-squares}
\keyword{tsls}
\keyword{two-stage}
